Extract framework for encapsulating drawable object data.

+++ cosmos/src/com/aimlessblade/test3d/Color.java
@@ -6,14 +6,14 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class Color {
-    public static final Color BLUE = new Color(0.0f, 0.0f, 1.0f);
-    public static final Color GRAY = new Color(0.8f, 0.8f, 0.8f);
-    public static final Color GREEN = new Color(0.0f, 1.0f, 0.0f);
-    public static final Color YELLOW = new Color(0.5f, 0.5f, 0.0f);
-    public static final Color RED = new Color(1.0f, 0.0f, 0.0f);
-    public static final Color CYAN = new Color(0.0f, 1.0f, 1.0f);
+    public static final Color BLUE = new Color(0.0, 0.0, 1.0);
+    public static final Color GRAY = new Color(0.8, 0.8, 0.8);
+    public static final Color GREEN = new Color(0.0, 1.0, 0.0);
+    public static final Color YELLOW = new Color(0.5, 0.5, 0.0);
+    public static final Color RED = new Color(1.0, 0.0, 0.0);
+    public static final Color CYAN = new Color(0.0, 1.0, 1.0);
     
-    private final float red;
-    private final float green;
-    private final float blue;
+    private final double red;
+    private final double green;
+    private final double blue;
 }
